add Template.class.ts
